+++ Source/union-miyoomini-toolchain/workspace/retrodex/core/activities/menu/input.cc
@@ -1,0 +1,109 @@
+#include "PokedexActivityManager.h"
+#include "PokedexActivityMenu.h"
+
+void PokedexActivityMenu::onButtonUp(SDL_Keycode sym, Uint16 mod)
+{
+    needRedraw = true;
+
+    if (selectedIndex > 0)
+    {
+        selectedIndex--;
+        if (selectedIndex < offset)
+        {
+            offset--;
+        }
+        // Play the sound effect
+        Mix_PlayChannel(1, se_up_down, 0);
+    }
+}
+
+void PokedexActivityMenu::onButtonDown(SDL_Keycode sym, Uint16 mod)
+{
+    needRedraw = true;
+
+    if (selectedIndex < dbResults->size() - 1)
+    {
+        selectedIndex++;
+        if (selectedIndex - offset >= MAX_VISIBLE_ITEMS)
+        {
+            offset++;
+        }
+        // Play the sound effect
+        Mix_PlayChannel(1, se_up_down, 0);
+    }
+}
+
+void PokedexActivityMenu::onButtonLeft(SDL_Keycode sym, Uint16 mod) {}
+
+void PokedexActivityMenu::onButtonRight(SDL_Keycode sym, Uint16 mod) {}
+
+void PokedexActivityMenu::onButtonA(SDL_Keycode sym, Uint16 mod)
+{
+    PokedexDB::setVersionID(std::stoi(game[0]));
+    PokedexDB::setGenerationID(std::stoi(game[5]));
+    PokedexDB::setVersionGroupID(std::stoi(game[7]));
+
+    PokedexActivityManager::push(APPSTATE_POKEDEX_LIST);
+}
+
+void PokedexActivityMenu::onButtonB(SDL_Keycode sym, Uint16 mod) {}
+
+void PokedexActivityMenu::onButtonR(SDL_Keycode sym, Uint16 mod)
+{
+    needRedraw = true;
+
+    if (selectedIndex < dbResults->size() - 3)
+    {
+        selectedIndex += 3;
+        if (selectedIndex - offset >= MAX_VISIBLE_ITEMS)
+        {
+            offset += 3;
+            // Ensure offset doesn't go out of bounds
+            if (offset > dbResults->size() - MAX_VISIBLE_ITEMS)
+            {
+                offset = dbResults->size() - MAX_VISIBLE_ITEMS; // Cap to last visible items
+            }
+        }
+        // Play the sound effect
+        Mix_PlayChannel(1, se_up_down, 0);
+    }
+    else
+    {
+        // If we exceed the last item, set selectedIndex to the last item visible
+        selectedIndex = dbResults->size() - 1;
+        offset        = dbResults->size() - MAX_VISIBLE_ITEMS; // Cap to last visible items
+    }
+}
+
+void PokedexActivityMenu::onButtonL(SDL_Keycode sym, Uint16 mod)
+{
+    needRedraw = true;
+
+    if (selectedIndex >= 3)
+    {
+        selectedIndex -= 3;
+        if (selectedIndex < offset)
+        {
+            offset -= 3; // Reduce offset accordingly
+            // Ensure offset doesn't go negative
+            if (offset < 0)
+            {
+                offset = 0; // Cap offset to zero
+            }
+        }
+        // Play the sound effect
+        Mix_PlayChannel(1, se_up_down, 0);
+    }
+    else
+    {
+        selectedIndex = 0; // Ensure selectedIndex doesn't go below zero
+        offset        = 0; // Cap offset to zero
+    }
+}
+
+void PokedexActivityMenu::onButtonSelect(SDL_Keycode sym, Uint16 mod) {}
+
+void PokedexActivityMenu::onButtonStart(SDL_Keycode sym, Uint16 mod)
+{
+    PokedexActivityManager::push(APPSTATE_POKEDEX_SETTING);
+}
