+++ Source/union-miyoomini-toolchain/workspace/retrodex/core/activities/pokemon_moves/render.cc
@@ -1,27 +1,5 @@
 #include "PokedexActivity_PokemonView_Moves.h"
-#include "PokedexActivityManager.h"
-#include "utils/name_to_id.h"
-
-PokedexActivity_PokemonView_Moves PokedexActivity_PokemonView_Moves::instance;
-
-PokedexActivity_PokemonView_Moves::PokedexActivity_PokemonView_Moves()
-    : dbResults(nullptr), pokemon(nullptr), needRedraw(true), backgroundSurface(nullptr),
-      listEntrySurface(nullptr), pokeIcon(nullptr), pokeName(nullptr), typeA(nullptr),
-      typeB(nullptr), se_left_right(nullptr), se_up_down(nullptr), fontSurface(nullptr),
-      selectedIndex(0), offset(0)
-{
-}
-
-PokedexActivity_PokemonView_Moves::~PokedexActivity_PokemonView_Moves()
-{
-    if (se_up_down)
-        Mix_FreeChunk(se_up_down);
-    se_up_down = nullptr;
-
-    if (se_left_right)
-        Mix_FreeChunk(se_left_right);
-    se_left_right = nullptr;
-}
+#include "name_to_id.h"
 
 bool PokedexActivity_PokemonView_Moves::initSDL()
 {
@@ -116,56 +94,6 @@ bool PokedexActivity_PokemonView_Moves::initSDL()
 
     return true;
 }
-void PokedexActivity_PokemonView_Moves::print_dbResults()
-{
-    for (std::vector<std::string> &moves : *dbResults)
-    {
-        for (auto &col : moves)
-        {
-            std::cout << col << " | ";
-        }
-        std::cout << std::endl;
-    }
-}
-
-void PokedexActivity_PokemonView_Moves::onActivate()
-{
-    std::cout << "PokedexActivity_PokemonView_Moves::onActivate START \n";
-
-    pokemon   = new Pokemon();
-    dbResults = PokedexDB::executeSQL(&SQL_getPokeMovesDetail);
-    print_dbResults();
-
-    move = (*dbResults)[selectedIndex];
-
-    assetManager = AssetManager::getInstance();
-
-    if (!initSDL())
-    {
-        std::cout
-            << "PokedexActivity_PokemonView_Moves::onActivate - Error in initSDL(), SDL Error: "
-            << std::endl;
-        exit(EXIT_FAILURE);
-    }
-
-    needRedraw = true;
-
-    std::cout << "PokedexActivity_PokemonView_Moves::onActivate END \n";
-}
-
-void PokedexActivity_PokemonView_Moves::onDeactivate()
-{
-    std::cout << "PokedexActivity_PokemonView_Moves::onDeactivate START \n";
-
-    delete pokemon;
-    pokemon = nullptr;
-
-    selectedIndex = 0, offset = 0;
-
-    std::cout << "PokedexActivity_PokemonView_Moves::onDeactivate END \n";
-}
-
-void PokedexActivity_PokemonView_Moves::onLoop() {}
 
 void PokedexActivity_PokemonView_Moves::onRender(SDL_Surface *surf_display, SDL_Renderer *renderer,
                                                  SDL_Texture *texture, TTF_Font *font,
@@ -339,144 +267,4 @@ bool PokedexActivity_PokemonView_Moves::renderItemDetails(SDL_Surface *surf_disp
     SDL_BlitSurface(summary, NULL, surf_display, &summaryRect);
 
     return true;
-}
-
-std::string PokedexActivity_PokemonView_Moves::cleanString(std::string str)
-{
-    std::string replacement = " ";
-    std::string cleanStr    = str;
-    size_t pos              = 0;
-    while ((pos = cleanStr.find("\n", pos)) != std::string::npos)
-    {
-        cleanStr.replace(pos, 1, replacement);
-        pos += replacement.length();
-    }
-
-    pos = 0;
-    while ((pos = cleanStr.find("\r", pos)) != std::string::npos)
-    {
-        cleanStr.replace(pos, 1, replacement);
-        pos += replacement.length();
-    }
-
-    pos = 0;
-    while ((pos = cleanStr.find("\f", pos)) != std::string::npos)
-    {
-        cleanStr.replace(pos, 1, replacement);
-        pos += replacement.length();
-    }
-
-    return cleanStr;
-}
-
-void PokedexActivity_PokemonView_Moves::onFreeze()
-{
-    // do thing for now..
-}
-
-PokedexActivity_PokemonView_Moves *PokedexActivity_PokemonView_Moves::getInstance()
-{
-    return &instance;
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonUp(SDL_Keycode sym, Uint16 mod)
-{
-    needRedraw = true;
-
-    if (selectedIndex > 0)
-    {
-        selectedIndex--;
-        if (selectedIndex < offset)
-        {
-            offset--;
-        }
-        Mix_PlayChannel(1, se_up_down, 0);
-    }
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonDown(SDL_Keycode sym, Uint16 mod)
-{
-    needRedraw = true;
-
-    if (selectedIndex < dbResults->size() - 1)
-    {
-        selectedIndex++;
-        if (selectedIndex - offset >= MAX_VISIBLE_ITEMS)
-        {
-            offset++;
-        }
-        Mix_PlayChannel(1, se_up_down, 0);
-    }
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonR(SDL_Keycode sym, Uint16 mod)
-{
-    needRedraw = true;
-
-    if (selectedIndex < dbResults->size() - MAX_VISIBLE_ITEMS)
-    {
-        selectedIndex += MAX_VISIBLE_ITEMS;
-        if (selectedIndex - offset >= MAX_VISIBLE_ITEMS)
-        {
-            offset += MAX_VISIBLE_ITEMS;
-            if (offset > dbResults->size() - MAX_VISIBLE_ITEMS)
-            {
-                offset = dbResults->size() - MAX_VISIBLE_ITEMS;
-            }
-        }
-    }
-    else
-    {
-        selectedIndex = dbResults->size() - 1;
-        offset        = dbResults->size() - MAX_VISIBLE_ITEMS;
-    }
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonL(SDL_Keycode sym, Uint16 mod)
-{
-    needRedraw = true;
-
-    if (selectedIndex >= MAX_VISIBLE_ITEMS)
-    {
-        selectedIndex -= MAX_VISIBLE_ITEMS;
-        if (selectedIndex < offset)
-        {
-            offset -= MAX_VISIBLE_ITEMS;
-            if (offset < 0)
-            {
-                offset = 0;
-            }
-        }
-    }
-    else
-    {
-        selectedIndex = 0;
-        offset        = 0;
-    }
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonLeft(SDL_Keycode sym, Uint16 mod)
-{
-    Mix_PlayChannel(1, se_left_right, 0);
-    PokedexActivityManager::replace(APPSTATE_POKEMON_VIEW_STATS);
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonRight(SDL_Keycode sym, Uint16 mod)
-{
-    Mix_PlayChannel(1, se_left_right, 0);
-    PokedexActivityManager::replace(APPSTATE_POKEMON_VIEW_LOCATION);
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonA(SDL_Keycode sym, Uint16 mod) {}
-
-void PokedexActivity_PokemonView_Moves::onButtonB(SDL_Keycode sym, Uint16 mod)
-{
-    PokedexActivityManager::back();
-}
-
-void PokedexActivity_PokemonView_Moves::onButtonSelect(SDL_Keycode sym, Uint16 mod) {}
-
-void PokedexActivity_PokemonView_Moves::onButtonStart(SDL_Keycode sym, Uint16 mod)
-{
-    PokedexActivityManager::push(APPSTATE_POKEDEX_SETTING);
 }
